+++ src/Symfony/Component/Lock/LockFactory.php
@@ -43,7 +43,19 @@ class LockFactory implements LoggerAwareInterface
      */
     public function createLock(string $resource, ?float $ttl = 300.0, bool $autoRelease = true): LockInterface
     {
-        $lock = new Lock(new Key($resource), $this->store, $ttl, $autoRelease);
+        return $this->createLockFromKey(new Key($resource), $ttl, $autoRelease);
+    }
+
+    /**
+     * Creates a lock from the given key.
+     *
+     * @param Key        $key         The key containing the lock's state
+     * @param float|null $ttl         Maximum expected lock duration in seconds
+     * @param bool       $autoRelease Whether to automatically release the lock or not when the lock instance is destroyed
+     */
+    public function createLockFromKey(Key $key, ?float $ttl = 300.0, bool $autoRelease = true): LockInterface
+    {
+        $lock = new Lock($key, $this->store, $ttl, $autoRelease);
         $lock->setLogger($this->logger);
 
         return $lock;